Feat: implement ProductRepository with Single Source of Truth:
-Added ProductRepositoryImpl with local-first strategy.

-Implemented background refresh logic using Mutex and Flow.

-Added ProductEntity and Mappers for Remote-Entity-Domain conversion.

-Integrated Local and Remote data sources.

+++ app/src/main/java/com/ccandeladev/androidtesting/productlist/data/remote/response/InventoryResponse.kt
@@ -18,7 +18,7 @@ data class ProductResponse(
     @SerialName("description")
     val description: String? = null,
     @SerialName("priceCents")
-    val priceCents: Int? = null,
+    val priceCents: Double? = null,
     @SerialName("category")
     val category: String? = null,
     @SerialName("stock")
